\name{Msimulated}
\docType{data}
\alias{Msimulated}
\title{Matrix of simulated ordinal data}
\description{
 This is a toy dataset for running simple examples.
}
\usage{Msimulated}

\format{ An ordinal data matrix with 60 lines and 50 columns. The number of levels is equal to 3. Four blocks are simulated with (mu,pi) parameters equal to (3,0.5), (2,0.7), (1,0.8) and (2,0.6).
}
%%\references{
%% McNeil, D. R. (1977) \emph{Interactive Data Analysis}.
%% New York: Wiley.
%%}
\keyword{datsaets}