\name{plot}
\docType{methods}
\alias{plot}
\alias{plot,ResultClassifOrdinal-method}
\alias{plot,ResultClustOrdinal-method}
\alias{plot,ResultCoclustOrdinal-method}
\title{ ~~ Method for the function \code{plot} in the \pkg{ordinalClust} package ~~}
\description{
Plots the result of a classification, clustering or co-clustering performed using the following functions:
bosclassif,bosclust,boscoclust.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "ResultClassifOrdinal")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "ResultClustOrdinal")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "ResultCoclustOrdinal")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
