\name{predict}
\docType{methods}
\alias{predict}
\alias{predict,ANY-method}
\alias{predict,ResultClassifOrdinal-method}
\title{ ~~ Method for the function \code{predict} in the \pkg{ordinalClust} package  ~~}
\description{
Method for the function \code{predict} in the \pkg{ordinalClust} package.
}
\section{Methods}{
\describe{


\item{\code{signature(object = "ResultClassifOrdinal")}}{
%%  ~~describe this method here~~
Use this method with the result of the function bosclassif, and a new sample to predict the classes.
}
}}
\keyword{methods}
